fix(ci-tests): the integration harness comes back under the file-length cap (#2163)
Not this branch's file, and unblocking it is why it is here: go-file-length
runs over the WHOLE tree, so harness.go at 505 lines fails every open PR
regardless of what that PR touches. origin/main is red on it today (#2163);
PR #2154 carried it over while adding a partner fixture.

Split along the grain rather than at whatever line the cap fell on: the row
fixtures — seed a row, answer its id — move to harnessseed.go, and Env
construction, the principal builders and the workspace-bound exec helpers stay
put. A suite reaches for exactly one of those three at a time.

The waiver file was not the answer. Its own header says it holds files that
were ALREADY over when the gate landed, frozen at their count, and that
shrinking is allowed while growing is not — a newly-over file belongs on the
other side of that line.
